restructure app dir and align design of prototype components
