+++ ui/app/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { Habit } from "../lib/interfaces";
-import { HabitCard } from "./habits/habitcard";
-import { NewHabit } from "./habits/NewHabit";
+import { HabitCard } from "./(habits)/HabitCard";
+import { NewHabit } from "./(habits)/NewHabit";
 
 const sampleHabits: Habit[] = [
   {
